fixed Errors, useAlert and Font
Errors:
backend/app.js - shifted app.use(errorMiddleware) to just above the if statement for process.env.NODE_ENV = "PRODUCTION" because it was blocking the errorMiddleware from being read.

useAlert()
useAlert uses context. So the topmost component (index.js) needs to be wrapped with <AlertProvider> in order for useAlert() to be possible to use.

Font
font URL for Amazon Ember was http so not secured and heroku was preventing the app from reading from the URL. switched to Nunito from Google fonts.